refactor: remove license, cleanup
